Split journal questions by day: feeling → today, alcohol → yesterday
Semantics:
- Today's journal: work/off + how did you feel after waking up?
- Yesterday's journal: did you drink alcohol? + free-text summary
- Intake: just supplements

Implementation:
- WorkDayToggle renamed to TodayJournal — now handles both the work/off
  chips (already there) and the morning-feeling radios. Both fields auto-
  save; each save round-trips the full journal entry so preserved-through
  fields (supplements, alcohol, notes) stay intact.
- IntakeLog loses the alcohol free-text input. It keeps drank_alcohol +
  alcohol_amount as preserved-through fields so its "Save supplements"
  button (renamed) doesn't wipe alcohol set on the Journal tab.
- JournalPage loses the morning-feeling fieldset and gains the alcohol
  free-text input. morning_feeling + is_work_day + followed_supplements
  are now the preserved-through fields on this form.
- DailyPage: "Supplements & alcohol" section relabelled "Supplements".
  TodayJournal replaces WorkDayToggle in the Today section.
