+++ frontend/src/components/DailyPage.tsx
@@ -5,8 +5,8 @@ import { IntakeLog } from "./IntakeLog";
 import { JournalPage } from "./JournalPage";
 import { NutritionTodayCard } from "./NutritionTodayCard";
 import { OodaPage } from "./OodaPage";
+import { TodayJournal } from "./TodayJournal";
 import { WaterQuickLog } from "./WaterQuickLog";
-import { WorkDayToggle } from "./WorkDayToggle";
 
 const today = format(new Date(), "yyyy-MM-dd");
 const yesterday = format(subDays(new Date(), 1), "yyyy-MM-dd");
@@ -15,8 +15,8 @@ export function DailyPage() {
   return (
     <OodaPage
       sections={[
-        { id: "today", label: "Today", content: <WorkDayToggle date={today} /> },
-        { id: "intake", label: "Supplements & alcohol", content: <IntakeLog /> },
+        { id: "today", label: "Today", content: <TodayJournal date={today} /> },
+        { id: "intake", label: "Supplements", content: <IntakeLog /> },
         { id: "journal", label: "Yesterday's journal", content: <JournalPage initialDate={yesterday} /> },
         { id: "water", label: "Water", content: <WaterQuickLog date={today} /> },
         { id: "activity", label: "Activities & steps", content: <AutoTickedToday date={today} /> },
